src/ocrinvoice/cli/utils.py
"""
CLI utilities for Invoice OCR Parser.

This module contains utility functions for the CLI including
logging setup and configuration loading.
"""

import logging
import yaml
from pathlib import Path
from typing import Dict, Any, Optional

logger = logging.getLogger(__name__)


def setup_logging(level: str = "INFO", log_file: Optional[str] = None) -> None:
    """
    Set up logging configuration.

    Args:
        level: Logging level
        log_file: Optional log file path
    """
    # Placeholder implementation
    logging.basicConfig(
        level=getattr(logging, level.upper()),
        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    )

    if log_file:
        file_handler = logging.FileHandler(log_file)
        file_handler.setFormatter(
            logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
        )
        logging.getLogger().addHandler(file_handler)

    logger.info("Logging configured")


def load_config(config_file: Optional[str] = None) -> Dict[str, Any]:
    """
    Load configuration from file.

    Args:
        config_file: Path to configuration file

    Returns:
        Configuration dictionary
    """
    # Placeholder implementation
    default_config = {
        "ocr": {"tesseract_path": None, "language": "eng", "timeout": 30},
        "parser": {"confidence_threshold": 0.7, "debug_mode": False},
        "output": {"format": "json", "include_confidence": True},
    }

    if config_file and Path(config_file).exists():
        try:
            with open(config_file, "r") as f:
                config = yaml.safe_load(f)
                logger.info(f"Configuration loaded from: {config_file}")
                return config
        except Exception as e:
            logger.error(f"Error loading config from {config_file}: {e}")

    logger.info("Using default configuration")
    return default_config


def validate_pdf_file(filepath: str) -> bool:
    """
    Validate that file is a PDF.

    Args:
        filepath: Path to file

    Returns:
        True if file is a valid PDF
    """
    # Placeholder implementation
    path = Path(filepath)
    if not path.exists():
        logger.error(f"File not found: {filepath}")
        return False

    if not path.suffix.lower() == ".pdf":
        logger.error(f"File is not a PDF: {filepath}")
        return False

    logger.info(f"PDF file validated: {filepath}")
    return True


def create_output_directory(output_path: str) -> bool:
    """
    Create output directory if it doesn't exist.

    Args:
        output_path: Output directory path

    Returns:
        True if directory created or exists
    """
    # Placeholder implementation
    path = Path(output_path)
    try:
        path.mkdir(parents=True, exist_ok=True)
        logger.info(f"Output directory ready: {output_path}")
        return True
    except OSError as e:
        logger.error(f"Error creating output directory {output_path}: {e}")
        return False
